requirements.txt
Django>=4.0,<6.0
djangorestframework>=3.14.0
python-decouple>=3.8